+++ client/src/pages/LoginPage.jsx
@@ -1,0 +1,3 @@
+export default function LoginPage() {
+  return <div>Login Page now routed</div>
+}
