+++ client/src/pages/RegisterPage.jsx
@@ -1,0 +1,3 @@
+export default function RegisterPage() {
+  return <div>Register page now routed</div>
+}
